chore: :zap: moving files and directories where they need to go using architecture
